requirements/base.txt
Django==4.1.2
django-environ==0.9.0
psycopg2==2.9.3
djangorestframework==3.13.1
drf-spectacular==0.22.1

whitenoise==6.2.0

django-filter==22.1
django-extensions==3.2.0
django-cors-headers==3.13.0
django-storages==1.12.3

drf-jwt==1.19.2

boto3==1.24.92
attrs==21.4.0

requests==2.28.1

wagtail==4.0.4

phonenumbers==8.12.53
django-countries==7.3.2
django-phonenumber-field==6.3.0
